Checkpoint file, status corrections, +rebrand
