+++ src/Rpc/HttpClient.h
@@ -23,6 +23,7 @@
 #include <HTTP/HttpResponse.h>
 #include <System/TcpConnection.h>
 #include <System/TcpStream.h>
+#include "JsonRpc.h"
 
 #include "Serialization/SerializationTools.h"
